Unpin the roster search and mail compose fields too
The guild roster search box had the same clipping as the create prompt,
from a separate rule: .guild-search pinned to 20px, tighter than the
dialog's 26px. Its parent .guild-filter-bar was pinned as well, so the bar
would have stayed a ceiling on the field even after the field itself
stopped asking for a fixed height -- both are min-height now.

The first version of this fixture should have caught it and did not, which
is the more useful half of this commit. The sweep looked for style rules
whose SELECTOR mentioned an input, and a field named for its job wears no
such name: .guild-search, .mail-compose__field. Rewritten to start from the
UXML and ask what a TextField is actually wearing, it immediately found two
more pinned fields in mail compose -- To and Subject at 22px, quantity and
currency at 22px.

Verified by restoring each original value and confirming the sweep reports
it, rather than by trusting a green run: all four are caught, and the tree
is clean with the fixes in place.

Multiline fields stay exempt. A text area is a sized, scrolling box, which
is a different control -- the map note body and the mail message body are
both legitimately fixed, and .mail-compose__body already used min-height.
