--- FishMMO-Unity/Assets/UnitTests/DialogInputHeightTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c198a6f4b5b1b4749afef70d00417314
